kingdom-manager/requirements.txt
tkinter
pillow
requests
